Move model-related data generators to a subpackage
