+++ src/main/java/net/minecraftforge/client/model/generators/BlockstateProvider.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package net.minecraftforge.client.model;
+package net.minecraftforge.client.model.generators;
 
 import com.google.common.base.Preconditions;
 import com.google.gson.Gson;
