+++ src/main/java/net/minecraftforge/client/model/generators/ModelProvider.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package net.minecraftforge.client.model;
+package net.minecraftforge.client.model.generators;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -25,7 +25,7 @@ import net.minecraft.data.DataGenerator;
 import net.minecraft.data.DirectoryCache;
 import net.minecraft.data.IDataProvider;
 import net.minecraft.util.ResourceLocation;
-import net.minecraftforge.client.model.ModelFile.GeneratedModelFile;
+import net.minecraftforge.client.model.generators.ModelFile.GeneratedModelFile;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
